Refactor grammar interpreter into function
