+++ packages/langium/src/index.ts
@@ -13,7 +13,6 @@ export * from './generator/node-processor';
 export * from './generator/template-string';
 export * from './grammar/generated/ast';
 export * from './grammar/generated/module';
-export * from './grammar/grammar-access';
 export * from './grammar/grammar-util';
 export * from './grammar/langium-grammar-module';
 export * from './grammar/language-meta-data';
@@ -23,7 +22,7 @@ export * from './lsp/completion/rule-interpreter';
 export * from './lsp/language-server';
 export * from './validation/document-validator';
 export * from './validation/validation-registry';
-export * from './parser/langium-base-parser';
+export * from './parser/langium-parser';
 export * from './references/linker';
 export * from './references/naming';
 export * from './references/scope';
